+++ Intrinio.Collections/RingBuffers/PartitionedRoundRobinRingBuffer.cs
@@ -14,7 +14,7 @@ public class PartitionedRoundRobinRingBuffer : IPartitionedRingBuffer
     private ulong _readIndex;
     private readonly SingleProducerRingBuffer[] _queues;
     private readonly uint _blockSize;
-    private readonly uint _eachBlockCapacity;
+    private readonly ulong _eachBlockCapacity;
     private readonly ulong _totalBlockCapacity;
     
     /// <summary>
@@ -25,7 +25,7 @@ public class PartitionedRoundRobinRingBuffer : IPartitionedRingBuffer
     /// <summary>
     /// The fixed capacity of blocks in each ring buffer.
     /// </summary>
-    public uint EachQueueBlockCapacity { get { return _eachBlockCapacity; } }
+    public ulong EachQueueBlockCapacity { get { return _eachBlockCapacity; } }
     
     /// <summary>
     /// The fixed total capacity of blocks across all ring buffers.
@@ -100,7 +100,7 @@ public class PartitionedRoundRobinRingBuffer : IPartitionedRingBuffer
     /// <param name="blockSize">The fixed size of each byte block.</param>
     /// <param name="eachQueueBlockCapacity">The fixed capacity of block count in each channel.</param>
     /// <exception cref="ArgumentException">Throws an ArgumentException if concurrency is zero.</exception>
-    public PartitionedRoundRobinRingBuffer(uint concurrency, uint blockSize, uint eachQueueBlockCapacity)
+    public PartitionedRoundRobinRingBuffer(uint concurrency, uint blockSize, ulong eachQueueBlockCapacity)
     {
         if (concurrency == 0U)
             throw new ArgumentException("Argument concurrency must be greater than zero.", nameof(concurrency));
